fixing warnings by converting values to int (line number: because we'll never get files with more than 2^32-1 lines)

+++ include/Ark/Compiler/Lexer.inl
@@ -77,7 +77,7 @@ inline void Lexer::throwTokenizingError(const std::string& message, const std::s
     {
         int iline = static_cast<int>(line);
 
-        if (iline - i >= 0 && iline - i < ctx.size())
+        if (iline - i >= 0 && iline - i < static_cast<int>(ctx.size()))  // cast to int, we'll never have a size > 2^32-1
             // + 1 to display real lines numbers
             ss << std::setw(5) << (iline - i + 1) << " | " << ctx[iline - i] << "\n";
         if (i == 0)  // line of the error
